[TASK] Remove service, create new functions and adapt call
The translateCopyService was removed and replaced
by two new functions in the Generator class.
The General class calling the service is adapted to
the new functions
