+++ Classes/TcaDataGenerator/TableHandler/General.php
@@ -17,10 +17,10 @@ namespace TYPO3\CMS\Styleguide\TcaDataGenerator\TableHandler;
 
 use TYPO3\CMS\Core\Database\ConnectionPool;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
+use TYPO3\CMS\Styleguide\TcaDataGenerator\Generator;
 use TYPO3\CMS\Styleguide\TcaDataGenerator\RecordData;
 use TYPO3\CMS\Styleguide\TcaDataGenerator\RecordFinder;
 use TYPO3\CMS\Styleguide\TcaDataGenerator\TableHandlerInterface;
-use TYPO3\CMS\Styleguide\Service\TranslateCopyService;
 
 /**
  * General table handler
@@ -66,28 +66,28 @@ class General extends AbstractTableHandler implements TableHandlerInterface
             [ 'uid' => $fieldValues['uid'] ]
         );
 
-        $translateCopyService = GeneralUtility::makeInstance(TranslateCopyService::class);
+        $localizeFunctions = GeneralUtility::makeInstance(Generator::class);
         $demoLanguages = $recordFinder->findUidsOfDemoLanguages();
 
         $translatedRecord = -42;
         foreach ($demoLanguages as $demoLanguageIndex => $demoLanguageUid) {
             switch ($demoLanguageIndex) {
                 case 0:
-                    $translateCopyService->copyRecordToLanguage($tableName, $fieldValues['uid'], $demoLanguageUid);
+                    $localizeFunctions->copyRecordToLanguage($tableName, $fieldValues['uid'], $demoLanguageUid);
                     break;
                 case 1:
-                    $result = $translateCopyService->localizeRecord($tableName, $fieldValues['uid'], $demoLanguageUid);
+                    $result = $localizeFunctions->localizeRecord($tableName, $fieldValues['uid'], $demoLanguageUid);
                     $translatedRecord = $result[$tableName][$fieldValues['uid']];
                     break;
                 case 2:
-                    $result = $translateCopyService->localizeRecord($tableName, $translatedRecord, $demoLanguageUid);
+                    $result = $localizeFunctions->localizeRecord($tableName, $translatedRecord, $demoLanguageUid);
                     $translatedRecord = $result[$tableName][$translatedRecord];
                     break;
                 case 3:
-                    $translateCopyService->copyRecordToLanguage($tableName, $translatedRecord, $demoLanguageUid);
+                    $localizeFunctions->copyRecordToLanguage($tableName, $translatedRecord, $demoLanguageUid);
                     break;
                 default:
-                    $translateCopyService->localizeRecord($tableName, $fieldValues['uid'], $demoLanguageUid);
+                    $localizeFunctions->localizeRecord($tableName, $fieldValues['uid'], $demoLanguageUid);
                     break;
             }
         }
